Removed UnWanted Lines

+++ server/controllers/userController.js
@@ -4,7 +4,6 @@ const fs = require("fs");
 const path = require("path");
 
 module.exports.update = async (req, res) => {
-  // console.log("Updating");
   let response = {
     success: false,
     message: "",
@@ -12,7 +11,6 @@ module.exports.update = async (req, res) => {
   };
   try {
     const { name, description, email, oldImage } = req.body;
-    console.log(req.body);
     let image;
     if (req.file) {
       temp = req.file.filename.split(".");
@@ -26,10 +24,6 @@ module.exports.update = async (req, res) => {
       { email: email },
       {
         $set: {
-          // image: {
-          //   data: fs.readFileSync("public/images/UserPic/" + req.file.filename),
-          //   contentType: "image/png",
-          // },
           image,
           name,
           description,
@@ -38,15 +32,11 @@ module.exports.update = async (req, res) => {
       { new: true }
     );
     res.json({ message: "User updated successfully", user });
-    // oldImage != "" &&
-    if (image != oldImage) {
+    if (oldImage != "" && image != oldImage) {
       let imageName = oldImage.split("/");
-      console.log(imageName);
-      console.log(imageName[imageName.length - 1]);
       let imagePath =
         path.join(__dirname, "../public/images/UserPic/") +
         imageName[imageName.length - 1];
-      console.log(imagePath);
       fs.unlink(imagePath, (err) => {
         if (err) {
           response.errMessage = err.message;
@@ -64,8 +54,6 @@ module.exports.update = async (req, res) => {
 };
 
 module.exports.getProfile = async (req, res) => {
-  // console.log("Profile");
-  console.log(req.user);
   let response = {
     success: true,
     message: "",
@@ -74,22 +62,11 @@ module.exports.getProfile = async (req, res) => {
   };
   let { id } = req.params;
   try {
-    await User.findOne({ _id: id })
-      // .populate([{ select: "name description image" }])
-      .then((result) => {
-        // console.log(result);
-        // let imageName = result.image.split("/");
-        // console.log(imageName);
-        // console.log(imageName[imageName.length - 1]);
-        // let imagePath =
-        //   path.join(__dirname, "../public/images/UserPic/") +
-        //   imageName[imageName.length - 1];
-        // console.log(imagePath);
-        // result.image = imagePath;
-        response.result = result;
-        response.success = true;
-        res.status(200).json(response);
-      });
+    await User.findOne({ _id: id }).then((result) => {
+      response.result = result;
+      response.success = true;
+      res.status(200).json(response);
+    });
   } catch (err) {
     console.log("Error", err);
     response.message = "Something went wrong!";
@@ -128,7 +105,6 @@ module.exports.addWallet = async (req, res) => {
     message: "",
     errMessage: "",
   };
-  console.log(req.body, "Dhiraj");
   const { id, wallet } = req.body;
   try {
     await User.findOneAndUpdate(
